# simplify sort+unique functions