--- a/maths/points/sorted_test.go
+++ b/maths/points/sorted_test.go
@@ -35,6 +35,10 @@ func TestSortAndUnique(t *testing.T) {
 			spts:  []maths.Pt{{1, 2}},
 		},
 		tcase{
+			uspts: []maths.Pt{{1, 2}, {1, 2}, {3, 4}, {5, 6}, {5, 6}},
+			spts:  []maths.Pt{{1, 2}, {3, 4}, {5, 6}},
+		},
+		tcase{
 			uspts: []maths.Pt{{7, 8}, {1, 2}, {3, 4}, {5, 6}, {3, 4}, {1, 2}, {7, 8}, {2, 3}, {1, 2}},
 			spts:  []maths.Pt{{1, 2}, {2, 3}, {3, 4}, {5, 6}, {7, 8}},
 		},