previously, the "find unique" part of these functions would shift everything down when it found unique value(s). This has a worst case runtime of O(n^2). New approach shifts down as you go, O(n). 

While this will probably have little impact on real world performance I found the old version difficult to follow with the 2 extra if conditions.

benchmarks are with 400ish floats, the sort is included as part of the benchmarks:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkSortUniqueF64-4               16916         15703         -7.17%
BenchmarkSortUniqueF64_worstCase-4     20154         14002         -30.52%
```