@gdey 2018-02-27
Thanks for the pull request. I will review it soon. Any improvements to the algo speed will help, as this function is called on every feature.